.claude/skills/brightness-contrast/scripts/brightness_contrast.py
#!/usr/bin/env python3
"""
GIMP 3.x Python-Fu script: Brightness/Contrast
Adjusts image brightness and contrast.

Config variables at the top are injected by Claude before writing to a temp file.
"""

# === CONFIG (injected by Claude) ===
INPUT_PATH = "__INPUT_PATH__"
OUTPUT_PATH = "__OUTPUT_PATH__"
BRIGHTNESS = __BRIGHTNESS__    # -1.0 to 1.0
CONTRAST = __CONTRAST__        # -1.0 to 1.0

# === IMPORTS ===
import gi
gi.require_version("Gimp", "3.0")
gi.require_version("Gegl", "0.4")
from gi.repository import Gimp, Gegl, Gio


def run():
    image = Gimp.file_load(Gimp.RunMode.NONINTERACTIVE, Gio.File.new_for_path(INPUT_PATH))
    if image is None:
        print(f"ERROR: Failed to load image: {INPUT_PATH}")
        return

    drawable = image.get_layers()[0]
    print(f"Loaded: {drawable.get_width()}x{drawable.get_height()}")

    drawable.brightness_contrast(BRIGHTNESS, CONTRAST)
    print(f"Applied: brightness={BRIGHTNESS}, contrast={CONTRAST}")

    image.flatten()
    Gimp.file_save(Gimp.RunMode.NONINTERACTIVE, image, Gio.File.new_for_path(OUTPUT_PATH))
    print(f"Output saved to: {OUTPUT_PATH}")
    image.delete()

run()
