requirements.txt
# Claude API (chat agent)
anthropic>=0.40.0

# PyTorch (requires ARM Python 3.12 on Apple Silicon)
torch>=2.4

# HTTP client (for Ollama API)
httpx>=0.27.0

# Configuration
pydantic>=2.0
pydantic-settings>=2.0

# YOLO object detection
ultralytics>=8.0

# CLIP semantic understanding
transformers>=4.36.0
Pillow>=10.0
sentencepiece>=0.1.99
